feat: Default to u32 when maximum is unset
Few object-store parameters set the format int64 and minimum to 0.
Latest changes in the format handling result in u32 params in the
generated code. Change the login to default at u32 when format is not
specified, but u64 when int64 with min is specified. Additionally
improve mentioned object-store schemas.


Changes are triggered by https://review.opendev.org/960413

+++ openstack_types/src/object_store/v1/account/response/get.rs
@@ -31,7 +31,7 @@ pub struct AccountResponse {
     /// The number of objects in the container.
     #[serde(default)]
     #[structable(optional)]
-    pub count: Option<u32>,
+    pub count: Option<u64>,
 
     /// Last modification date of the container
     #[serde(default)]
